Actually, the authentication class DID work. I had just forgot to give user testapp the rights to add log messages. Moves authentication to new file and restructured the code.
